shift funcs in linalg.py
